Try lower-cased filename
Filenames of DLLs are sometimes lower case on mingw, while the library name
is uppercase. This is OK on Windows, but easily breaks on Unix-like systems
that are case sensitive. However, Windows DLLs are almost always all lower
case on mingw.

Try to find a lower cased variant of the library in addition to the file name
listed as "DLL Name" to workaround.